requirements.txt
ddgs
requests
beautifulsoup4
selenium
scholarly
python-dotenv